Make Gaussian wrapping GSObjectModel (#38)
