+++ piff/__init__.py
@@ -75,9 +75,7 @@ from .config import piffify, setup_logger, read_config
 # Models -- Class names here match what they are called in the config file
 from .model import Model
 from .pixelgrid import PixelGrid, Lanczos, Bilinear
-from .gaussian_model import Gaussian
-from .kolmogorov_model import Kolmogorov
-from .gsobject_model import GSObjectModel
+from .gsobject_model import GSObjectModel, Gaussian
 
 # Interpolators -- Class names here match what they are called in the config file
 from .interp import Interp
