+++ piff/util.py
@@ -140,3 +140,32 @@ def read_kwargs(fits, extname):
     assert len(data) == 1
     kwargs = dict([ (col, data[col][0]) for col in cols ])
     return kwargs
+
+def hsm(star):
+    """ Use HSM to measure moments of star image.
+    """
+    import galsim
+    image, weight, image_pos = star.data.getImage()
+    mom = image.FindAdaptiveMom(weight=weight, strict=False)
+
+    sigma = mom.moments_sigma
+    shape = mom.observed_shape
+    # These are in pixel coordinates.  Need to convert to world coords.
+    jac = image.wcs.jacobian(image_pos=image_pos)
+    scale, shear, theta, flip = jac.getDecomposition()
+    # Fix sigma
+    sigma *= scale
+    # Fix shear.  First the flip, if any.
+    if flip:
+        shape = galsim.Shear(g1 = -shape.g1, g2 = shape.g2)
+    # Next the rotation
+    shape = galsim.Shear(g = shape.g, beta = shape.beta + theta)
+    # Finally the shear
+    shape = shear + shape
+
+    flux = mom.moments_amp
+
+    center = image.wcs.toWorld(mom.moments_centroid) - image.wcs.toWorld(image_pos)
+    flag = mom.moments_status
+
+    return flux, center.x, center.y, sigma, shape.g1, shape.g2, flag
